# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TRADE FAST GOLD Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: List all active investment plans
      responses:
        "200":
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createPlan
      tags: [plans]
      summary: Create investment plan (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Created plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /plans/{id}:
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get investment plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    patch:
      operationId: updatePlan
      tags: [plans]
      summary: Update investment plan (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanUpdate"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: deletePlan
      tags: [plans]
      summary: Delete investment plan (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── INVESTMENTS ──────────────────────────────────────────────────────────
...
  /kyc:
    get:
      operationId: getKyc
      tags: [kyc]
      summary: Get KYC status for current user
      responses:
        "200":
          description: KYC record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycRecord"
    post:
      operationId: submitKyc
      tags: [kyc]
      summary: Submit KYC documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycInput"
      responses:
        "201":
          description: KYC submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycRecord"

  # ─── REFERRALS ────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List user notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /support/tickets:
    get:
      operationId: listTickets
      tags: [support]
      summary: List support tickets
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
    post:
      operationId: createTicket
      tags: [support]
      summary: Create support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Created ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
  /support/tickets/{id}:
    get:
      operationId: getTicket
      tags: [support]
      summary: Get support ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicketDetail"
  /support/tickets/{id}/reply:
    post:
      operationId: replyTicket
      tags: [support]
      summary: Reply to support ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketReplyInput"
      responses:
        "201":
          description: Reply sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketMessage"

  # ─── ADMIN ────────────────────────────────────────────────────────────────
...
  /admin/kyc:
    get:
      operationId: listAdminKyc
      tags: [admin]
      summary: List KYC submissions (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: KYC submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminKycPage"
  /admin/kyc/{id}/approve:
    patch:
      operationId: approveKyc
      tags: [admin]
      summary: Approve KYC (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved KYC
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycRecord"
  /admin/kyc/{id}/reject:
    patch:
      operationId: rejectKyc
      tags: [admin]
      summary: Reject KYC (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected KYC
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycRecord"
  /admin/tickets:
    get:
      operationId: listAdminTickets
      tags: [admin]
      summary: List all support tickets (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Paginated tickets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTicketPage"
components:
  schemas:
...
    Plan:
      type: object
      required: [id, name, tier, minAmount, maxAmount, roiPercent, durationDays, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        tier:
          type: string
          enum: [starter, silver, gold, premium, vip]
        description:
          type: ["string", "null"]
        minAmount:
          type: number
        maxAmount:
          type: number
        roiPercent:
          type: number
        durationDays:
          type: integer
        isActive:
          type: boolean
        features:
          type: array
          items:
            type: string
        createdAt:
          type: string
    PlanInput:
      type: object
      required: [name, tier, minAmount, maxAmount, roiPercent, durationDays]
      properties:
        name:
          type: string
        tier:
          type: string
          enum: [starter, silver, gold, premium, vip]
        description:
          type: string
        minAmount:
          type: number
        maxAmount:
          type: number
        roiPercent:
          type: number
        durationDays:
          type: integer
        features:
          type: array
          items:
            type: string
    PlanUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        minAmount:
          type: number
        maxAmount:
          type: number
        roiPercent:
          type: number
        durationDays:
          type: integer
        isActive:
          type: boolean
        features:
          type: array
          items:
            type: string

    # ─── INVESTMENTS ────────────────────────────────────────────────────────
...
    KycRecord:
      type: object
      required: [id, userId, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected, resubmit_required]
        documentType:
          type: ["string", "null"]
        documentFrontUrl:
          type: ["string", "null"]
        documentBackUrl:
          type: ["string", "null"]
        selfieUrl:
          type: ["string", "null"]
        proofOfAddressUrl:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        reviewedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    KycInput:
      type: object
      required: [documentType, documentFrontUrl]
      properties:
        documentType:
          type: string
          enum: [passport, drivers_license, national_id]
        documentFrontUrl:
          type: string
        documentBackUrl:
          type: string
        selfieUrl:
          type: string
        proofOfAddressUrl:
          type: string

    # ─── REFERRALS ──────────────────────────────────────────────────────────
...
    Notification:
      type: object
      required: [id, userId, type, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, investment, profit, referral, kyc, security, system]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
...
    SupportTicket:
      type: object
      required: [id, userId, subject, category, status, priority, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        subject:
          type: string
        category:
          type: string
          enum: [deposit, withdrawal, investment, account, technical, other]
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        createdAt:
          type: string
        updatedAt:
          type: string
    SupportTicketDetail:
      type: object
      required: [id, userId, subject, category, status, priority, messages, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        subject:
          type: string
        category:
          type: string
        status:
          type: string
        priority:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/TicketMessage"
        createdAt:
          type: string
    TicketMessage:
      type: object
      required: [id, ticketId, senderRole, message, createdAt]
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        senderRole:
          type: string
          enum: [user, admin]
        message:
          type: string
        createdAt:
          type: string
    TicketInput:
      type: object
      required: [subject, category, message]
      properties:
        subject:
          type: string
        category:
          type: string
          enum: [deposit, withdrawal, investment, account, technical, other]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        message:
          type: string
    TicketReplyInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
    RejectInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string

    # ─── ADMIN ──────────────────────────────────────────────────────────────
...
    AdminKycPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/KycRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminTicketPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SupportTicket"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer