# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TRADE FAST GOLD Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanTier } from './planTier';

export interface Plan {
  id: number;
  name: string;
  tier: PlanTier;
  /** @nullable */
  description?: string | null;
  minAmount: number;
  maxAmount: number;
  roiPercent: number;
  durationDays: number;
  isActive: boolean;
  features?: string[];
  createdAt: string;
}