# lib/api-zod/src/generated/types/planUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TRADE FAST GOLD Investment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PlanUpdate {
  name?: string;
  description?: string;
  minAmount?: number;
  maxAmount?: number;
  roiPercent?: number;
  durationDays?: number;
  isActive?: boolean;
  features?: string[];
}