# lib/api-zod/src/generated/types/supportTicketDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TRADE FAST GOLD Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TicketMessage } from './ticketMessage';

export interface SupportTicketDetail {
  id: number;
  userId: number;
  subject: string;
  category: string;
  status: string;
  priority: string;
  messages: TicketMessage[];
  createdAt: string;
}